fakeh_news/requirements.txt
aioboto3==15.1.0
aiobotocore==2.24.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.5
boto3==1.39.11
botocore==1.39.11
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
deepframe==0.1.1
deeplake==3.9.52
dill==0.4.0
discord.py==2.6.3
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
executing==2.2.1
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humbug==0.3.2
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
lz4==4.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.4
multiprocess==0.70.18
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
parso==0.8.5
pathos==0.3.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.4.0
pox==0.3.6
ppft==1.7.7
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-telegram-bot==22.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
s3transfer==0.13.1
sentry-sdk==2.37.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.47.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
